Centralizing creation of nonmem models and processing only the discrete model file passed by the user. No inferences in file level calculations
